requirements.txt
Django==1.8.8
Mezzanine==4.0.1
Pillow==3.1.0
argparse==1.2.1
beautifulsoup4==4.4.1
bleach==1.4.2
chardet==2.3.0
django-contrib-comments==1.6.2
django-webbugger==0.1
filebrowser-safe==0.4.0
future==0.15.2
grappelli-safe==0.4.1
gunicorn==19.4.5
html5lib==0.9999999
oauthlib==1.0.3
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
tzlocal==1.2
wsgiref==0.1.2
